Backend test suite: tests for models and API   controllers

+++ Backend/app/controllers/api/wishlists_controller.rb
@@ -34,6 +34,8 @@ module Api
     end
 
     def remove_item
+      raise ActiveRecord::RecordNotFound if current_user.wishlist.nil?
+
       wishlist_item = current_user.wishlist.wishlist_items.find(params[:id])
       wishlist_item.destroy!
